Call verify! on adapter instance before running a query
Since Active Record 7.1, `verify!` is no longer called on the connection when checking it out from the pool.
`verify!` is the method that calls `configure_connection`, which in turn sets up the type map for PostgreSQL.

This causes a race condition where if you try to execute a query on that instance before `verify!` gets called down the line, you get this error:

```
lib/active_record/connection_adapters/postgresql_adapter.rb in get_oid_type
undefined method `key?' for nil:NilClass (NoMethodError)
if !type_map.key?(oid)
```

`verify!` uses a monitor to synchronize calls so it is safe to call it here prior to running a query.
So, in case of multiple threads using the same adapter instance (which only happens in the context of transactional fixtures in test environment),
it is a thread safe operation.

Custom gems or extensions that rely on checking out a connection from the pool must now first call `verify!` on it if they need `configure_connection` to be called.